Fix player skip bug
If a player removed a player before them in the turn order the next
player was being skipped, because of the global turn counter being used.
Bug e.g.
Players: [p1 p2 p3 p4]
On p2's turn (turn 2 globally) p1 is removed the turn counter would go
up to 3 which when compared to the size of the array means p3 would be
skipped and p4 would go next.